LIBAVALON-158. Fix transitive Jena inclusion of JDK library
The "jena-core" library includes the "xml-apis" package as a
transitive dependency. This is technically an error as packages
included in the JDK should not be included as dependencies.

This transitive dependency causes Eclipse to show many errors of the
form:

```
The package javax.xml.stream is accessible from more than one module: <unnamed>, java.xml
```

Added a Maven "exclusion" rule to exclude the "xml-apis" dependency on
the "jena-core" package. This fixes the errors in Eclipse, and should
have no effect on the runtime behavior.

https://issues.umd.edu/browse/LIBAVALON-158