LIBAVALON-158. Refactored MigratorFedora2Export to be more testable
Refactored  MigratorFedora2Export to be more testable by extracting
a "createProcessor" method that could be overridden by the tests.
This was needed to enable the appropriate directory holding the
test data to be provided.

Changed the visibility of some of the instance variables so that
they would be accessible by subclasses.

Created "MigratorFedora2ExportTest" to test the MigratorFedora2Export
class. This test class creates a subclass of MigratorFedora2Export
that overrides the "createProcessor" method.

Added sample input data for use in the export, as well as the expected "export.csv" file and other outputs.

https://issues.umd.edu/browse/LIBAVALON-158