LIBAVALON-167. Modified title parsing to only use title under "descMeta"
The XML being processed may have multiple XML elements with a "title"
tag. The "umdm_title" provided in the JSON should only be derived from
"title" tags that are direct children of the "descMeta" tag -- all
other title tags should be ignored.

Modified the "UmdTitleParser" class to track the parent tags using
a stack-like structure, and only process a title tag if the "descMeta"
tag is the immediate parent.

Updated expected "info.json" to have the correct expected title.
https://issues.umd.edu/browse/LIBAVALON-167