LIBAVALON-167. Simplified code and added JavaDoc comments
Changed "title" instance variable to "titles" List, which allows
String.join in "getTitle" to concatenate the titles with the delimiter.

Also removed a null check for "title_type" by reversing the string
"equals" comparison.

Renamed UmdmTitleParser to "UmdmTitleXmlEventParser" to make it a
little clearer that it needs to be called for each event from
the XMLStreamReader.

Added JavaDoc.

https://issues.umd.edu/browse/LIBAVALON-167