LIBAVALON-166. Convert bibref to multiline general note
Modified archival collection "bibref" tags into a multiline general
note. This code assumes that the bibref consists of a "title" tag
and one or more "<bibscope>" tags, where the "<bibscope>" tags have
a "type" attribute that should be displayed as a caption.

Added a script providing simple unit tests.

Added "scripts/__pycache__" to ".gitignore" file, as they should not
be checked in.

https://issues.umd.edu/browse/LIBAVALON-166