LIBAVALON-166. Refactored code and added "ordering" constraints to output
Refactored the bibref conversion code into a "BibRefToTextConverter"
class, as the code was getting fairly complicated, and wanted to clarify
the code by organizing it into its own class.

As part of that refactoring, moved the "collapse_whitespace_nodes" and
"get_text" utility methods into an "XmlUtils" class. This way, the
"BibRefToTextConverter" class is dependent on another class, not the
functions in the script.

https://issues.umd.edu/browse/LIBAVALON-166