LIBAVALON-172. Refactored code to make it more testable.
Added "CsvColumnCounts" class to calculate and manage the "multicolumn"
field counts in the CSV output.

Added "ObjectToCsvConverter" class to manage the CSV output. This
places the "headers" generation, and the CSV output into the same
class, which seemed advantageous, since they would need to be modified
together if there are changes. Also moved the "multicolumn" function
into this class as it is used for the CSV output.

Modified the "write_csv" function to used the refactored classes.

Wrote some unit tests to verify the behavior of the new classes.

https://issues.umd.edu/browse/LIBAVALON-172