LIBHYDRA-586. Fixed "creator/contributor" parsing in "archelon.py"
Modified the "creator" and "contributor" parsing in the “process_umdm”
method of “scripts/archelon.py” to accept both `None` and the empty
string ("") for determining "agent_role" field is empty. Prior to this
change the "agent_role" was considered empty only if it was `None`.

Bumped the patch version of the "PyYAML" requirement from "6.0" to
"6.0.1", so the dependency could be used (otherwise was attempting to
use Cython 3.0, which is incompatible).

https://umd-dit.atlassian.net/browse/LIBHYDRA-586